Declare memory displacement/scale as long to trigger out of bounds values
  Otherwise, the compiler may just silently truncate bad values, and
the new code for 32 bit range check would become useless.

+++ lightning/i386/asm-64.h
@@ -308,35 +308,40 @@ _REXFQrr(jit_fpr_t rr, jit_gpr_t mr)
 
 /* --- ALU instructions ---------------------------------------------------- */
 __jit_inline void
-_ALUQrr(x86_alu_t op, jit_gpr_t rs, jit_gpr_t rd)
+_ALUQrr(x86_alu_t op,
+	jit_gpr_t rs, jit_gpr_t rd)
 {
     _REXQrr(rs, rd);
     _alu_sil_rr(op, rs, rd);
 }
 
 __jit_inline void
-_ALUQmr(x86_alu_t op, int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+_ALUQmr(x86_alu_t op,
+	long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _alu_sil_mr(op, md, rb, ri, ms, rd);
 }
 
 __jit_inline void
-_ALUQrm(x86_alu_t op, jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+_ALUQrm(x86_alu_t op,
+	jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _alu_sil_rm(op, rs, md, rb, ri, ms);
 }
 
 __jit_inline void
-_ALUQir(x86_alu_t op, long im, jit_gpr_t rd)
+_ALUQir(x86_alu_t op,
+	long im, jit_gpr_t rd)
 {
     _REXQrr(_NOREG, rd);
     _alu_il_ir(op, im, rd);
 }
 
 __jit_inline void
-_ALUQim(x86_alu_t op, long im, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+_ALUQim(x86_alu_t op,
+	long im, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(_NOREG, rb, ri);
     _alu_il_im(op, im, md, rb, ri, ms);
@@ -392,21 +397,24 @@ _ALUQim(x86_alu_t op, long im, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
 
 /* --- Shift/Rotate instructions ------------------------------------------- */
 __jit_inline void
-_ROTSHIQir(x86_rotsh_t op, long im, jit_gpr_t rd)
+_ROTSHIQir(x86_rotsh_t op,
+	   long im, jit_gpr_t rd)
 {
     _REXQrr(_NOREG, rd);
     _rotsh_sil_ir(op, im, rd);
 }
 
 __jit_inline void
-_ROTSHIQim(x86_rotsh_t op, long im, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+_ROTSHIQim(x86_rotsh_t op,
+	   long im, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(_NOREG, rb, ri);
     _rotsh_sil_im(op, im, md, rb, ri, ms);
 }
 
 __jit_inline void
-_ROTSHIQrr(x86_rotsh_t op, jit_gpr_t rs, jit_gpr_t rd)
+_ROTSHIQrr(x86_rotsh_t op,
+	   jit_gpr_t rs, jit_gpr_t rd)
 {
     _REXQrr(rs, rd);
     _rotsh_sil_rr(op, rs, rd);
@@ -414,7 +422,7 @@ _ROTSHIQrr(x86_rotsh_t op, jit_gpr_t rs, jit_gpr_t rd)
 
 __jit_inline void
 _ROTSHIQrm(x86_rotsh_t op,
-	   jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+	   jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _rotsh_sil_rm(op, rs, md, rb, ri, ms);
@@ -462,28 +470,32 @@ _ROTSHIQrm(x86_rotsh_t op,
 
 /* --- Bit test instructions ----------------------------------------------- */
 __jit_inline void
-_BTQir(x86_bt_t op, long im, jit_gpr_t rd)
+_BTQir(x86_bt_t op,
+       long im, jit_gpr_t rd)
 {
     _REXQrr(_NOREG, rd);
     _bt_sil_ir(op, im, rd);
 }
 
 __jit_inline void
-_BTQim(x86_bt_t op, long im, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+_BTQim(x86_bt_t op,
+       long im, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(_NOREG, rb, ri);
     _bt_sil_im(op, im, md, rb, ri, ms);
 }
 
 __jit_inline void
-_BTQrr(x86_bt_t op, jit_gpr_t rs, jit_gpr_t rd)
+_BTQrr(x86_bt_t op,
+       jit_gpr_t rs, jit_gpr_t rd)
 {
     _REXQrr(rs, rd);
     _bt_sil_rr(op, rs, rd);
 }
 
 __jit_inline void
-_BTQrm(x86_bt_t op, jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+_BTQrm(x86_bt_t op,
+       jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _bt_sil_rm(op, rs, md, rb, ri, ms);
@@ -518,14 +530,14 @@ MOVQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-MOVQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+MOVQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _mov_sil_mr(md, rb, ri, ms, rd);
 }
 
 __jit_inline void
-MOVQrm(jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+MOVQrm(jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _mov_sil_rm(rs, md, rb, ri, ms);
@@ -540,7 +552,7 @@ MOVQir(long im, jit_gpr_t rd)
 }
 
 __jit_inline void
-MOVQim(long im, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+MOVQim(long im, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(_NOREG, rb, ri);
     _O(0xc7);
@@ -550,14 +562,16 @@ MOVQim(long im, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
 
 /* --- Unary and Multiply/Divide instructions ------------------------------ */
 __jit_inline void
-_UNARYQr(x86_unary_t op, jit_gpr_t rs)
+_UNARYQr(x86_unary_t op,
+	 jit_gpr_t rs)
 {
     _REXQrr(_NOREG, rs);
     _unary_sil_r(op, rs);
 }
 
 __jit_inline void
-_UNARYQm(x86_unary_t op, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+_UNARYQm(x86_unary_t op,
+	 long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQmr(rb, ri, _NOREG);
     _unary_sil_m(op, md, rb, ri, ms);
@@ -589,7 +603,7 @@ IMULQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-IMULQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+IMULQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _imul_sil_mr(md, rb, ri, ms, rd);
@@ -619,14 +633,16 @@ JMPQsr(jit_gpr_t rs)
 }
 
 __jit_inline void
-CMOVQrr(x86_cc_t cc,jit_gpr_t rs, jit_gpr_t rd)
+CMOVQrr(x86_cc_t cc,
+	jit_gpr_t rs, jit_gpr_t rd)
 {
     _REXQrr(rd, rs);
     _cmov_sil_rr(cc, rs, rd);
 }
 
 __jit_inline void
-CMOVQmr(x86_cc_t cc, int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+CMOVQmr(x86_cc_t cc,
+	long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _cmov_sil_mr(cc, md, rb, ri, ms, rd);
@@ -641,7 +657,7 @@ POPQr(jit_gpr_t rd)
 }
 
 __jit_inline void
-POPQm(int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+POPQm(long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQm(rb, ri);
     _pop_sil_m(md, rb, ri, ms);
@@ -655,7 +671,7 @@ PUSHQr(jit_gpr_t rs)
 }
 
 __jit_inline void
-PUSHQm(int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+PUSHQm(long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQm(rb, ri);
     _push_sil_m(md, rb, ri, ms);
@@ -676,7 +692,7 @@ TESTQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-TESTQrm(jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+TESTQrm(jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _test_sil_rm(rs, md, rb, ri, ms);
@@ -690,7 +706,7 @@ TESTQir(long im, jit_gpr_t rd)
 }
 
 __jit_inline void
-TESTQim(long im, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+TESTQim(long im, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(_NOREG, rb, ri);
     _test_il_im(im, md, rb, ri, ms);
@@ -705,7 +721,7 @@ CMPXCHGQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-CMPXCHGQrm(jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+CMPXCHGQrm(jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _cmpxchg_sil_rm(rs, md, rb, ri, ms);
@@ -719,7 +735,7 @@ XADDQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-XADDQrm(jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+XADDQrm(jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _xadd_sil_rm(rs, md, rb, ri, ms);
@@ -733,7 +749,7 @@ XCHGQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-XCHGQrm(jit_gpr_t rs, int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+XCHGQrm(jit_gpr_t rs, long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(rs, rb, ri);
     _xchg_sil_rm(rs, md, rb, ri, ms);
@@ -748,7 +764,7 @@ DECQr(jit_gpr_t rd)
 }
 
 __jit_inline void
-DECQm(int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+DECQm(long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(_NOREG, rb, ri);
     _dec_sil_m(md, rb, ri, ms);
@@ -762,7 +778,7 @@ INCQr(jit_gpr_t rd)
 }
 
 __jit_inline void
-INCQm(int md, jit_gpr_t rb, jit_gpr_t ri, int ms)
+INCQm(long md, jit_gpr_t rb, jit_gpr_t ri, long ms)
 {
     _REXQrm(_NOREG, rb, ri);
     _inc_sil_m(md, rb, ri, ms);
@@ -777,7 +793,7 @@ BSFQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-BSFQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+BSFQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _bsf_sil_mr(md, rb, ri, ms, rd);
@@ -791,7 +807,7 @@ BSRQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-BSRQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+BSRQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _bsr_sil_mr(md, rb, ri, ms, rd);
@@ -807,7 +823,7 @@ _movsd_l_rr(jit_gpr_t rs, jit_gpr_t rd)
 
 /* long rd = (int)*rs */
 __jit_inline void
-_movsd_l_mr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+_movsd_l_mr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _O(0x63);
     _r_X(rd, md, rb, ri, ms, 0);
@@ -821,7 +837,7 @@ MOVSBQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-MOVSBQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+MOVSBQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _movsb_sil_mr(md, rb, ri, ms, rd);
@@ -835,7 +851,7 @@ MOVSWQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-MOVSWQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+MOVSWQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _movsw_il_mr(md, rb, ri, ms, rd);
@@ -849,7 +865,7 @@ MOVSLQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-MOVSLQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+MOVSLQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _movsd_l_mr(md, rb, ri, ms, rd);
@@ -863,7 +879,7 @@ MOVZBQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-MOVZBQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+MOVZBQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _movzb_sil_mr(md, rb, ri, ms, rd);
@@ -877,14 +893,14 @@ MOVZWQrr(jit_gpr_t rs, jit_gpr_t rd)
 }
 
 __jit_inline void
-MOVZWQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+MOVZWQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _movzw_il_mr(md, rb, ri, ms, rd);
 }
 
 __jit_inline void
-LEAQmr(int md, jit_gpr_t rb, jit_gpr_t ri, int ms, jit_gpr_t rd)
+LEAQmr(long md, jit_gpr_t rb, jit_gpr_t ri, long ms, jit_gpr_t rd)
 {
     _REXQmr(rb, ri, rd);
     _lea_il_mr(md, rb, ri, ms, rd);
@@ -914,7 +930,6 @@ CQO_(void)
     _REXQrr(_NOREG, _NOREG);
     _sign_extend_rdx_rax();
 }
-
 
 #define __SSEQrr(OP, RS, RD)						\
     (_REXQrr(RD, RS),							\
